@@ -13,18 +13,20 @@
   "private": true,
   "dependencies": {
     "@agm/core": "^1.0.0-beta.2",
-    "@angular/animations": "^6.0.0-rc.5",
-    "@angular/cdk": "^6.0.0-rc.5",
-    "@angular/common": "^6.0.0-rc.5",
-    "@angular/compiler": "^6.0.0-rc.5",
-    "@angular/core": "^6.0.0-rc.5",
+    "@angular/animations": "^6.1.2",
+    "@angular/cdk": "^6.4.3",
+    "@angular/common": "^6.1.2",
+    "@angular/compiler": "^6.1.2",
+    "@angular/core": "^6.1.2",
     "@angular/flex-layout": "2.0.0-beta.10",
-    "@angular/forms": "^6.0.0-rc.5",
-    "@angular/http": "^6.0.0-rc.5",
-    "@angular/material": "^6.0.0-rc.5",
-    "@angular/platform-browser": "^6.0.0-rc.5",
-    "@angular/platform-browser-dynamic": "^6.0.0-rc.5",
-    "@angular/router": "^6.0.0-rc.5",
+    "@angular/forms": "^6.1.2",
+    "@angular/http": "^6.1.2",
+    "@angular/material": "^6.4.3",
+    "@angular/platform-browser": "^6.1.2",
+    "@angular/platform-browser-dynamic": "^6.1.2",
+    "@angular/router": "^6.1.2",
+    "@ngx-translate/core": "^10.0.2",
+    "@ngx-translate/http-loader": "^3.0.1",
     "@swimlane/ngx-charts": "^7.0.1",
     "@swimlane/ngx-datatable": "11.1.5",
     "angular-calendar": "0.23.3",
@@ -35,11 +37,14 @@
     "d3": "^4.9.1",
     "dragula": "3.7.2",
     "font-awesome": "4.7.0",
+    "fullcalendar": "^3.6.1",
     "hammerjs": "2.0.8",
     "intl": "^1.2.4",
     "jquery": "3.2.1",
+    "karma-jasmine": "^1.1.1",
     "leaflet": "^1.0.2",
     "moment": "^2.15.0",
+    "ng-fullcalendar": "^1.6.2",
     "ng2-charts": "1.6.0",
     "ng2-dragula": "1.3.1",
     "ng2-file-upload": "1.2.1",
@@ -49,7 +54,6 @@
     "ngx-perfect-scrollbar": "5.0.0-7",
     "perfect-scrollbar": "^1.3.0",
     "quill": "^1.2.0",
-    "karma-jasmine": "^1.1.1",
     "rxjs": "^6.0.0-uncanny-rc.7",
     "rxjs-compat": "^6.0.0-uncanny-rc.7",
     "screenfull": "^3.2.2",
@@ -60,20 +64,20 @@
   },
   "devDependencies": {
     "@angular-devkit/build-angular": "~0.6.1",
-    "@angular/cli": "^6.0.0-rc.5",
-    "@angular/compiler-cli": "^6.0.0-rc.5",
-    "@types/node": "~8.9.4",
-    "codelyzer": "~4.1.0",
+    "@angular/cli": "^6.1.3",
+    "@angular/compiler-cli": "^6.1.2",
     "@types/jasmine": "2.5.52",
     "@types/jquery": "3.2.0",
+    "@types/node": "~8.9.4",
+    "codelyzer": "~4.1.0",
     "jasmine-core": "2.6.3",
     "jasmine-spec-reporter": "4.1.0",
     "karma": "1.7.0",
     "karma-chrome-launcher": "2.1.1",
     "karma-cli": "1.0.1",
+    "karma-coverage-istanbul-reporter": "1.3.0",
     "karma-jasmine": "1.1.0",
     "karma-jasmine-html-reporter": "^0.2.2",
-    "karma-coverage-istanbul-reporter": "1.3.0",
     "protractor": "5.1.2",
     "ts-node": "~5.0.0",
     "tslint": "~5.9.1",
