@@ -1,18 +1,18 @@
 import {Injectable} from '@angular/core';
 import {Observable} from 'rxjs/Observable';
-import { Http } from '@angular/http';
+import { HttpClient } from '@angular/common/http';
 
 const TOKEN_KEY = 'AuthToken';
 
 @Injectable()
 export class AuthService {
 
-    constructor(private http: Http) { }
+    constructor(private http: HttpClient) { }
   
     attemptAuth(username: string, password: string): Observable<any> {
       const credentials = {username: username, password: password};
       console.log('attempAuth ::');
-      return this.http.post('http://localhost:8080/login', credentials);
+      return this.http.post('http://localhost:8080/login', credentials, {responseType:"text"});
     }
 
     signOut() {
