@@ -1,21 +1,23 @@
 import { NgModule } from '@angular/core';
+import { FormsModule } from '@angular/forms';
 import { ListComponent } from './list/list.component';
 import { EditComponent } from './edit/edit.component';
 import { RouterModule } from '@angular/router';
-import { PacienteRoutes } from './paciente.routing';
-import { PacienteService } from './paciente.service';
+import { ConsultaService } from './consulta.service';
+import { ConsultaRoutes } from './consulta.routing';
 import { SharedModule } from '../shared/shared.module';
 
 
 @NgModule({
   imports: [
-    RouterModule.forChild(PacienteRoutes),
+    RouterModule.forChild(ConsultaRoutes),
     SharedModule
   ],
   declarations: [ 
     ListComponent,
     EditComponent
-  ], providers: [PacienteService]
+  ], 
+  providers: [ConsultaService]
 })
 
-export class PacienteModule {}
+export class ConsultaModule {}
