@@ -1,11 +1,11 @@
 import { Injectable } from '@angular/core';
-import { HttpBaseService } from '../shared/httpbase.service';
 import { RestfulService } from '../shared/restful.service';
 import { Paciente } from './paciente';
+import { HttpClient } from '@angular/common/http';
 
 @Injectable()
 export class PacienteService extends RestfulService<Paciente> {
-    constructor( public http: HttpBaseService) {
+    constructor( public http: HttpClient) {
         super(http, 'public/paciente');
     }
 }
