@@ -2,6 +2,7 @@ import { Component, OnInit ,ViewEncapsulation } from '@angular/core';
 import {Router} from "@angular/router";
 import { MatDialog } from '@angular/material';
 import { AuthService } from '../../auth.service';
+import { HttpErrorResponse } from '@angular/common/http';
 
 @Component({
    selector: 'ms-login-session',
@@ -11,20 +12,25 @@ import { AuthService } from '../../auth.service';
 })
 export class LoginComponent {
   
-  constructor(private router: Router, public dialog: MatDialog, private authService: AuthService, private authservice: AuthService) { }
+  constructor(private router: Router,private authservice: AuthService) { }
 
   username: string;
   password: string;
 
   login(): void {
-      this.authService.attemptAuth(this.username, this.password).subscribe(
+      this.authservice.attemptAuth(this.username, this.password).subscribe(
       data => {
-        console.log(data._body);
-        this.authservice.saveToken(data._body);
+        console.log(data);
+        this.authservice.saveToken(data);
         this.router.navigate(['/home']);
       },
-      error => {
-        console.log("Erro ao tentar autenticar");
+      (err: HttpErrorResponse) => {
+        if (err.error instanceof Error) {
+          console.log("Client-side error occured.");
+        } else {
+          console.log("Server-side error occured.");
+        }
+        console.log(err.error);
       }
     );
   }	
